requirements.txt
customtkinter
psutil
py-cpuinfo
gputil
